requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.37
databases[sqlite]==0.9.0
aiosqlite==0.20.0
python-dotenv==1.0.1
jdatetime==5.2.0
requests
slowapi
newspaper3k
lxml-html-clean